# api/openapi.yaml
openapi: 3.0.3
info:
  title: Axiom owner-console release API
  version: 1.3.0
  description: >-
    Authenticated research, operations, audit, sandbox, and qualification
    control plane. Axiom V1 is Spot-only research software. Real trading,
    production-private order submission, and external asset movement are absent.
servers:
  - url: /
...
paths:
...
  /api/v1/system/status:
    get:
      operationId: getSystemStatus
      tags: [System]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Current fail-closed system projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/exchanges/binance/instruments:
    get:
      operationId: listBinanceInstruments
      tags: [Binance]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Deterministically ordered Spot metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstrumentPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/portfolios:
    get:
      operationId: listPortfolios
      tags: [Portfolio]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Virtual portfolio summaries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortfolioPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/portfolios/{id}:
    get:
      operationId: getPortfolio
      tags: [Portfolio]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Exact virtual balance and position projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortfolioDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/portfolios/{id}/journal:
    get:
      operationId: listPortfolioJournal
      tags: [Portfolio]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Exact immutable virtual journal page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JournalPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/risk/status:
    get:
      operationId: getRiskStatus
      tags: [Risk]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Effective global risk state and evidence
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/strategies/trend:
    get:
      operationId: getTrendStrategy
      tags: [Trend]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Immutable Trend version, parameters, and evidence maturity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrendStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/strategies/trend/decisions:
    get:
      operationId: listTrendDecisions
      tags: [Trend]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: outcome
          schema: { type: string, enum: [accepted, rejected] }
      responses:
        "200":
          description: Canonical decision and rejection explanations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrendDecisionPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/backtests/{id}:
    get:
      operationId: getBacktest
      tags: [Backtest]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Backtest lifecycle, result, and reproducibility projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResource" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/replays:
    post:
      operationId: createReplay
      tags: [Replay]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplayJobRequest" }
      responses:
        "202":
          description: Durable replay job accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/replays/{id}:
    get:
      operationId: getReplay
      tags: [Replay]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - in: query
          name: event_ordinal
          description: Select one persisted canonical replay event; defaults to the newest available event.
          schema: { $ref: "#/components/schemas/Revision" }
      responses:
        "200":
          description: Replay lifecycle and deterministic cursor projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResource" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/shadow-sessions:
    get:
      operationId: listShadowSessions
      tags: [Shadow]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: state
          schema:
            type: string
            enum: [QUEUED, RUNNING, PAUSED, CANCEL_REQUESTED, CANCELED, FAILED]
      responses:
        "200":
          description: Bounded deterministic public-data shadow history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShadowSessionPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createShadowSession
      tags: [Shadow]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShadowSessionRequest" }
      responses:
        "202":
          description: Public-data simulation-only shadow session accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShadowSessionResource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/shadow-sessions/{id}:
    get:
      operationId: getShadowSession
      tags: [Shadow]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Public-data virtual shadow projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShadowSessionResource" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/incidents:
    get:
      operationId: listIncidents
      tags: [Incident]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: state
          schema: { type: string, enum: [open, acknowledged, resolved] }
      responses:
        "200":
          description: Correlated incident summaries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncidentPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createIncident
      tags: [Incident]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IncidentCreateRequest" }
      responses:
        "202":
          description: Durable incident command accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandAccepted" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/incidents/{id}:
    get:
      operationId: getIncident
      tags: [Incident]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - in: query
          name: include_raw
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Correlated redacted incident timeline
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncidentDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/audit-events:
    get:
      operationId: listAuditEvents
      tags: [Audit]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: event_type
          schema: { type: string, maxLength: 80 }
        - in: query
          name: include_detail
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Redacted immutable audit page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditEventPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: axiom_session
    operationalBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Origin:
      in: header
      name: Origin
      required: true
      schema: { type: string, format: uri, maxLength: 512 }
    CSRFToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 32, maxLength: 512 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        {
          type: string,
          minLength: 16,
          maxLength: 128,
          pattern: "^[A-Za-z0-9._:-]+$",
        }
    ResourceID:
      in: path
      name: id
      required: true
      schema: { type: string, minLength: 1, maxLength: 160 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string, minLength: 16, maxLength: 1024 }
    PageSize:
      in: query
      name: page_size
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: Request failed safe validation
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Authentication failed
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authorization, Origin, or CSRF policy denied the request
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Idempotency payload or lifecycle conflict
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    PreconditionFailed:
      description: Expected entity revision is stale
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    RateLimited:
      description: Durable authentication or workload quota exceeded
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unavailable:
      description: A required fail-closed prerequisite is unavailable
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    NonnegativeDecimal:
      type: string
      pattern: "^(0|[1-9][0-9]*)(\\.[0-9]+)?$"
    Revision:
      type: string
      pattern: "^(0|[1-9][0-9]*)$"
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp with an explicit UTC offset.
...
    Page:
      type: object
      additionalProperties: false
      required: [revision, has_more]
      properties:
        revision: { $ref: "#/components/schemas/Revision" }
        next_cursor: { type: string, minLength: 16, maxLength: 1024 }
        has_more: { type: boolean }
...
    SystemStatus:
      type: object
      additionalProperties: false
      required:
        [
          application_version,
          build_commit,
          configuration_identity,
          readiness_state,
          lifecycle_state,
          strategy_activation,
          real_trading_enabled,
        ]
      properties:
        application_version: { type: string, minLength: 1 }
        build_commit: { type: string, minLength: 1 }
        configuration_identity: { type: string, minLength: 1 }
        readiness_state: { type: string, enum: [ready, blocked, degraded] }
        lifecycle_state:
          {
            type: string,
            enum: [STARTING, READY_PAUSED, RUNNING, DEGRADED, STOPPING],
          }
        strategy_activation:
          { type: string, enum: [unavailable, trend-following@1.0.0] }
        real_trading_enabled: { type: boolean, enum: [false] }
        environment: { type: string }
        execution_mode:
          { type: string, enum: [backtest, replay, paper, shadow] }
        engine_state: { type: string }
        binance_state: { type: string }
        risk_state: { type: string, enum: [PAUSED, RESUMED, LOCKED] }
        active_resource_id: { type: string }
        critical_incidents: { type: integer, minimum: 0 }
        server_time: { $ref: "#/components/schemas/Timestamp" }
        clock_drift_ms: { type: string, pattern: "^-?(0|[1-9][0-9]*)$" }
        revision: { $ref: "#/components/schemas/Revision" }
...
    InstrumentPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [items]
          properties:
            items:
              {
                type: array,
                items: { $ref: "#/components/schemas/Instrument" },
              }
...
    PortfolioPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [items]
          properties:
            items:
              {
                type: array,
                items: { $ref: "#/components/schemas/PortfolioSummary" },
              }
...
    PortfolioDetail:
      allOf:
        - $ref: "#/components/schemas/PortfolioSummary"
        - type: object
          required: [balances, positions, updated_at]
          properties:
            balances:
              { type: array, items: { $ref: "#/components/schemas/Balance" } }
            positions:
              { type: array, items: { $ref: "#/components/schemas/Position" } }
            updated_at: { $ref: "#/components/schemas/Timestamp" }
...
    JournalPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [virtual, items]
          properties:
            virtual: { type: boolean, enum: [true] }
            items:
              {
                type: array,
                items: { $ref: "#/components/schemas/JournalEntry" },
              }
...
    RiskStatus:
      type: object
      additionalProperties: false
      required:
        [
          state,
          revision,
          policy_version,
          contributors,
          recovery_ready,
          updated_at,
        ]
      properties:
        state: { type: string, enum: [NORMAL, CAUTIOUS, PAUSED, LOCKED] }
        revision: { $ref: "#/components/schemas/Revision" }
        policy_version: { $ref: "#/components/schemas/Revision" }
        reason_codes: { type: array, items: { type: string }, maxItems: 64 }
        contributors:
          {
            type: array,
            items: { $ref: "#/components/schemas/RiskContributor" },
          }
        recovery_ready: { type: boolean }
        unresolved_critical: { type: integer, minimum: 0 }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
...
    CommandAccepted:
      type: object
      additionalProperties: false
      required: [id, state, target_id, revision, correlation_id, created_at]
      properties:
        id: { type: string }
        state: { type: string, enum: [pending, applied, rejected, failed] }
        target_id: { type: string }
        revision: { $ref: "#/components/schemas/Revision" }
        correlation_id: { type: string }
        created_at: { $ref: "#/components/schemas/Timestamp" }
...
    TrendStatus:
      type: object
      additionalProperties: false
      required:
        [version, timeframe, health, evidence_maturity, parameters, revision]
      properties:
        version: { type: string, enum: [trend-following@1.0.0] }
        timeframe: { type: string, enum: [4h] }
        health:
          { type: string, enum: [healthy, warming, paused, degraded, locked] }
        evidence_maturity:
          {
            type: string,
            enum: [local_tier_b, formal_tier_a, insufficient, rejected],
          }
        viability:
          {
            type: string,
            enum: [undetermined, viable_for_more_research, rejected],
          }
        parameters:
          {
            type: array,
            minItems: 16,
            maxItems: 16,
            items: { $ref: "#/components/schemas/TrendParameter" },
          }
        revision: { $ref: "#/components/schemas/Revision" }
...
    TrendDecisionPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [items]
          properties:
            items:
              {
                type: array,
                items: { $ref: "#/components/schemas/TrendDecision" },
              }
...
    ReplayJobRequest:
      allOf:
        - $ref: "#/components/schemas/OfflineJobRequest"
        - type: object
          properties:
            speed: { type: string, enum: [original, accelerated, maximum] }
            incident_id: { type: string }
            first_ordinal: { $ref: "#/components/schemas/Revision" }
            last_ordinal: { $ref: "#/components/schemas/Revision" }
...
    JobResource:
      type: object
      additionalProperties: false
      required: [id, kind, state, mode_label, revision, created_at]
      properties:
        id: { type: string }
        kind: { type: string, enum: [backtest, replay] }
        state:
          {
            type: string,
            enum:
              [
                QUEUED,
                RUNNING,
                PAUSE_REQUESTED,
                PAUSED,
                CANCEL_REQUESTED,
                CANCELED,
                SUCCEEDED,
                FAILED,
              ],
          }
        mode_label: { type: string, enum: [BACKTEST, REPLAY] }
        revision: { $ref: "#/components/schemas/Revision" }
        progress: { $ref: "#/components/schemas/NonnegativeDecimal" }
        cursor_ordinal: { $ref: "#/components/schemas/Revision" }
        failure_code: { type: string }
        result: { $ref: "#/components/schemas/JobResult" }
        registered_report:
          { $ref: "#/components/schemas/RegisteredResearchReport" }
        replay_inspection:
          { $ref: "#/components/schemas/ReplayEventInspection" }
        checkpoints:
          type: array
          items: { $ref: "#/components/schemas/ReplayCheckpoint" }
        input_manifest: { $ref: "#/components/schemas/LabInputManifest" }
        lifecycle: { $ref: "#/components/schemas/LabLifecycleCapabilities" }
        reproduction_bundle: { $ref: "#/components/schemas/ReproductionBundle" }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
...
    ShadowSessionRequest:
      type: object
      additionalProperties: false
      required: [configuration_id, strategy_version, portfolio_id]
      properties:
        configuration_id: { type: string }
        strategy_version: { type: string, enum: [trend-following@1.0.0] }
        portfolio_id: { type: string }
...
    ShadowSessionPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/ShadowSessionSummary" }
...
    IncidentPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [items]
          properties:
            items:
              {
                type: array,
                items: { $ref: "#/components/schemas/IncidentSummary" },
              }
...
    IncidentCreateRequest:
      type: object
      additionalProperties: false
      required:
        [severity, reason_code, owner_user_id, expected_revision, reason]
      properties:
        severity: { type: string, enum: [warning, error, critical] }
        reason_code: { type: string, minLength: 3, maxLength: 120 }
        owner_user_id: { type: string, minLength: 1, maxLength: 160 }
        expected_revision: { $ref: "#/components/schemas/Revision" }
        reason: { type: string, minLength: 8, maxLength: 500 }
...
    AuditEventPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [items]
          properties:
            items:
              {
                type: array,
                items: { $ref: "#/components/schemas/AuditEvent" },
              }
...